rename: 헤더 관련 객체 이름 변경
